trace/obfuscator: refactor to simplify obfuscator config
Remove unnecessary duplicate structs:
* `obfuscate.Config` removed in favor of `config.ObfuscatorConfig`
* `obfuscate.JSONSettings` removed in favor of `config.JSONObfuscatorConfig`

Now it's no longer necessary to copy/convert from one struct to the other as was being done in the agent.

Motivation: Simplify obfuscator config & initialization as we are now using the obfuscator from python checks and need a way to use the same configuration as the obfuscator instance in the agent. See upcoming PR https://github.com/DataDog/datadog-agent/pull/5692.